Check DB constants are set by `$_ENV` when applicable
While an alternative would be to grep `wp-config.php` for use of
`$_ENV['DB_HOST']`, etc., the only way to truly check that DB constants
are set by `$_ENV` is to modify the launch check bootstrap process such
that the config checks happen after `wp-config.php` is loaded into
scope, but before the rest of WordPress is loaded. If the rest of
WordPress is loaded with invalid database credentials, then WP-CLI will
error.