Default “all” command to use --all
When this command runs on the platform, it does not explicitly utilize the --all flag, but that’s because the plugin command (which does have an option for --all) does not actually implement this logic - it will pull all plugins and makes no distinction whether they’re active or not.

This commit will just assume that when anyone (including the platform) runs the “all” command, it will expect the --all flag is enabled.